refactor: turned changelog generation into a pipe
I turned myself into a pipe morty!

PipeRick!!!

refs #284
