+++ internal/git/git.go
@@ -22,3 +22,8 @@ func Run(args ...string) (output string, err error) {
 	}
 	return string(bts), err
 }
+
+// Clean the output
+func Clean(output string, err error) (string, error) {
+	return strings.Replace(strings.Split(output, "\n")[0], "'", "", -1), err
+}
